# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Telegram Channel Manager API
  description: |
    # 🚀 Telegram Channel Manager v2.2
    
    سیستم حرفه‌ای مدیریت کانال‌های تلگرام با قابلیت‌های پیشرفته
    
    ## 🔑 احراز هویت
    تمام endpoint ها نیاز به کلید API دارند که باید در header ارسال شود:
    ```
    x-api-key: your_api_key_here
    ```
    
    ### دریافت کلید API
    1. با Master Key به `/api/keys/generate` درخواست POST بدید
    2. کلید API و Recovery Token رو ذخیره کنید
    3. در صورت گم شدن کلید، با Recovery Token می‌تونید بازیابی کنید
    
    ## ⚡ محدودیت‌های Rate Limit
    | عملیات | محدودیت | پنجره زمانی |
    |--------|----------|-------------|
    | Join Channel | 5 | دقیقه |
    | Leave Channel | 10 | دقیقه |
    | Channel Info | 30 | دقیقه |
    | Channel List | 10 | دقیقه |
    | Cleanup | 2 | ساعت |
    | Global | 100 | دقیقه |
    
    ## 📊 Session Management
    سشن‌ها از فایل `sessions.json` خوانده می‌شوند (مسیر قابل تنظیم در `SESSIONS_FILE_PATH`)
    
    ## 🔄 Response Codes
    - `200` - Success
    - `400` - Bad Request
    - `401` - Unauthorized (Invalid API Key)
    - `403` - Forbidden (No Permission)
    - `429` - Rate Limit Exceeded
    - `500` - Internal Server Error
    - `503` - Service Unavailable

  version: 2.2.0
  contact:
    name: Support Team
    email: support@telegram-manager.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development Server
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/session/auth/initialize:
    post:
      tags: [Session Auth]
      summary: Initialize Session Creation
      description: Start the process of creating a new Telegram session
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionName, phoneNumber]
              properties:
                sessionName:
                  type: string
                  example: "session_1"
                phoneNumber:
                  type: string
                  example: "+989121234567"
                isPremium:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Session initialized
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  sessionId:
                    type: string
                  message:
                    type: string
                  nextStep:
                    type: string
                    example: "requestCode"
  /api/session/auth/request-code:
    post:
      tags: [Session Auth]
      summary: Request Verification Code
      description: Request verification code to be sent to Telegram
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId]
              properties:
                sessionId:
                  type: string
      responses:
        '200':
          description: Code sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  nextStep:
                    type: string
                  codeLength:
                    type: integer
  /api/session/auth/submit-code:
    post:
      tags: [Session Auth]
      summary: Submit Verification Code
      description: Submit the verification code received from Telegram
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId, code]
              properties:
                sessionId:
                  type: string
                code:
                  type: string
                  example: "12345"
      responses:
        '200':
          description: Code verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  requiresPassword:
                    type: boolean
                  session:
                    type: object
  /api/session/auth/submit-password:
    post:
      tags: [Session Auth]
      summary: Submit 2FA Password
      description: Submit two-factor authentication password
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId, password]
              properties:
                sessionId:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Session created successfully
  /api/session/auth/cancel:
    post:
      tags: [Session Auth]
      summary: Cancel Session Creation
      description: Cancel the ongoing session creation process
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId]
              properties:
                sessionId:
                  type: string
      responses:
        '200':
          description: Session creation cancelled
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        API Key required for authentication
        
        Get your key:
        1. Use master key to call POST /api/keys/generate
        2. Save the API key and recovery token
        3. Use the API key in x-api-key header
...